# specs/005-core-domain-crud-apis/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AkarAI Phase 4 Core Domain CRUD API
  version: 0.1.0
  description: >
    Contract for Phase 4 core domain database and CRUD APIs. This contract excludes
    AI search, RAG, image upload/processing, OCR, email sending, dashboards,
    chatbot behavior, buyer-to-agency real-time chat, spam classification,
    lead scoring, and generated replies.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
  /agencies/me/profile:
    get:
      summary: Get current agency profile
      tags: [Agencies]
      responses:
        "200": { description: Agency profile }
    put:
      summary: Update current agency profile
      tags: [Agencies]
      responses:
        "200": { description: Updated agency profile }
        "403": { description: Support employees cannot update agency profile }
  /agencies/me/employees:
    get:
      summary: List agency employees
      tags: [Agencies]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated employees }
    post:
      summary: Create agency employee
      tags: [Agencies]
      responses:
        "201": { description: Employee created }
        "403": { description: Support employees cannot manage employees }
  /agencies/me/employees/{employee_id}:
    get:
      summary: Get agency employee
      tags: [Agencies]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200": { description: Employee }
    patch:
      summary: Update agency employee
      tags: [Agencies]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200": { description: Employee updated }
        "403": { description: Support employees cannot manage employees }
    delete:
      summary: Deactivate agency employee
      tags: [Agencies]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "204": { description: Employee deactivated }
...
  /listings/{listing_id}/inquiries:
    post:
      summary: Submit listing inquiry
      tags: [Leads]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "201": { description: Lead created }
        "429": { description: Inquiry rate limit exceeded }
  /agency/leads:
    get:
      summary: List agency leads
      tags: [Leads]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { description: Paginated tenant leads }
  /agency/leads/{lead_id}:
    get:
      summary: Get agency lead
      tags: [Leads]
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "200": { description: Lead }
    patch:
      summary: Update agency lead status
      tags: [Leads]
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "200": { description: Lead updated }
  /agency/leads/{lead_id}/review:
    post:
      summary: Mark lead reviewed
      tags: [Leads]
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "201": { description: Reviewed lead record created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    ListingId:
      name: listing_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    LeadId:
      name: lead_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    EmployeeId:
      name: employee_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...